Use a base class to keep timestamp sensor inheritance
