+++ homeassistant/components/jewish_calendar/entity.py
@@ -2,21 +2,15 @@
 
 from dataclasses import dataclass
 import datetime as dt
-import logging
 
 from hdate import HDateInfo, Location, Zmanim
 from hdate.translator import Language, set_language
 
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import SUN_EVENT_SUNSET
 from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
 from homeassistant.helpers.entity import Entity, EntityDescription
-from homeassistant.helpers.sun import get_astral_event_date
-from homeassistant.util import dt as dt_util
 
 from .const import DOMAIN
-
-_LOGGER = logging.getLogger(__name__)
 
 type JewishCalendarConfigEntry = ConfigEntry[JewishCalendarData]
 
@@ -70,43 +64,4 @@ class JewishCalendarEntity(Entity):
             location=self.data.location,
             candle_lighting_offset=self.data.candle_lighting_offset,
             havdalah_offset=self.data.havdalah_offset,
-        )
-
-    async def async_update_data(self) -> None:
-        """Update the state of the sensor."""
-        now = dt_util.now()
-        _LOGGER.debug("Now: %s Location: %r", now, self.data.location)
-
-        today = now.date()
-        event_date = get_astral_event_date(self.hass, SUN_EVENT_SUNSET, today)
-
-        if event_date is None:
-            _LOGGER.error("Can't get sunset event date for %s", today)
-            return
-
-        sunset = dt_util.as_local(event_date)
-
-        _LOGGER.debug("Now: %s Sunset: %s", now, sunset)
-
-        daytime_date = HDateInfo(today, diaspora=self.data.diaspora)
-
-        # The Jewish day starts after darkness (called "tzais") and finishes at
-        # sunset ("shkia"). The time in between is a gray area
-        # (aka "Bein Hashmashot"  # codespell:ignore
-        # - literally: "in between the sun and the moon").
-
-        # For some sensors, it is more interesting to consider the date to be
-        # tomorrow based on sunset ("shkia"), for others based on "tzais".
-        # Hence the following variables.
-        after_tzais_date = after_shkia_date = daytime_date
-        today_times = self.make_zmanim(today)
-
-        if now > sunset:
-            after_shkia_date = daytime_date.next_day
-
-        if today_times.havdalah and now > today_times.havdalah:
-            after_tzais_date = daytime_date.next_day
-
-        self.data.results = JewishCalendarDataResults(
-            daytime_date, after_shkia_date, after_tzais_date, today_times
         )
